Fix: marginError

+++ Statistics/marginError.py
@@ -5,7 +5,7 @@ from Calculator.squareRoot import squareRoot
 
 def MarginError(data):
     List = []
-    SE = int(StandardDeviationSample(data) / (squareRoot(len(data))))
+    SE = (StandardDeviationSample(data) / (squareRoot(len(data))))
     for i in Z_scores(z_values(data)):
         ME = i * SE
         List.append(ME)
